requirements.txt
argostranslategui-1.6.5